+++ packages/backend/package.json
@@ -2,7 +2,7 @@
   "name": "backend",
   "scripts": {
     "dev": "wrangler dev",
-    "build": "tsc",
+    "build": "tsc --noEmit",
     "deploy": "pnpm run build && wrangler deploy --minify",
     "cf-typegen": "wrangler types --env-interface CloudflareBindings",
     "test": "vitest",
