+++ packages/backend/src/infrastructure/repositories/DiscordTokensRepository.ts
@@ -1,6 +1,6 @@
 import { eq } from "drizzle-orm";
 import { injectable } from "inversify";
-import { db } from "../../../database/connection";
+import { db } from "../../../database/client";
 import { discordTokens as discordTokensSchema } from "../../../database/schema";
 import {
   AccessToken,
