+++ packages/backend/src/infrastructure/repositories/UserRepository.ts
@@ -1,6 +1,6 @@
 import { eq } from "drizzle-orm";
 import { injectable } from "inversify";
-import { db } from "../../../database/connection";
+import { db } from "../../../database/client";
 import { user as userSchema } from "../../../database/schema";
 import {
   Department,
